requirements.txt
tensorflow==2.1.0
streamlit==0.76.0
numpy==1.19.5
pillow==8.1.0
OpenCV :3.3.1
